pyproject.toml
[project]
name = "unilink-api"
version = "1.0.0"
description = "UniLink FastAPI Backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "motor>=3.3.2",
    "pymongo>=4.6.1",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "pydantic>=2.5.3",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "boto3>=1.34.34",
    "redis>=5.0.1",
    "aiosmtplib>=3.0.1",
    "httpx>=0.26.0",
    "google-generativeai>=0.3.2",
    "anthropic>=0.18.1",
    "authlib>=1.3.0",
    "itsdangerous>=2.1.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

# ADD THIS SECTION
[tool.hatch.build.targets.wheel]
packages = ["app"]